[refact] NoticeAlert 컴포넌트 갱신
