+++ src/components/common/NoticeAlert.jsx
@@ -1,69 +1,95 @@
-import React, { useCallback } from "react";
+import React, { PureComponent } from "react";
 import styled from "styled-components";
 import { Button } from "@material-ui/core";
-import { makeStyles } from "@material-ui/core/styles";
+import { styled  as mstyled } from '@material-ui/core/styles';
 // import warning from "./warning.svg";
 
-const useStyles = makeStyles((theme, color) => ({
-    leftBtn: {
-        flex: 1,
-        margin: "0 8px 0 8px",
-        borderRadius: "5px",
-        fontSize: "1em",
-        border: "1px solid #e5e8ec",
-        background: "#e5e8ec",
-    },
-    rightBtn: {
-        flex: 1,
-        margin: "0 8px 0 8px",
-        borderRadius: "5px",
-        fontSize: "1em",
-        border: "1px solid #e5e8ec",
-        background: "#e5e8ec",
-    },
-}));
+let that = null;
+class NoticeAlert extends PureComponent{
+    constructor(props){
+        super(props);
 
-const NoticeAlert = ({ icon, content, btns, handleOnClick, onCloseModal }) => {
-    const classes = useStyles();
+        this.state = {
+           open: false
+        };
 
-    return (
-        <>
-            <Modal>
-                <ModalCloser onClick={onCloseModal} />
-                <Section>
-                    <ModalMain>
-                        <NoticeContainer>
-                            <Notice>
-                                {/* <img src={warning} alt="warning" /> */}
-                                <div>{icon}</div>
-                                <div>{content}</div>
-                            </Notice>
-                        </NoticeContainer>
-                    </ModalMain>
-                    <ModalFooter>
-                        {btns.map((oneBtn) => {
-                            return (
-                                <Button
-                                    key={oneBtn.id}
-                                    className={
-                                        oneBtn.id == "left"
-                                            ? classes.leftBtn
-                                            : classes.rightBtn
-                                    }
-                                    onClick={() => handleOnClick(oneBtn.id)}
-                                >
-                                    {oneBtn.content}
-                                </Button>
-                            );
-                        })}
-                    </ModalFooter>
-                </Section>
-            </Modal>
-        </>
-    );
+        that = this;    //정적 메소드용
+    }
+    
+    static open (){
+        //open 여부를 소스 내에서 관리
+        if(!that.state.open)    that.setState({open: true});
+    }
+
+    handleOnClose = () => {
+        this.setState({open: false});
+    }
+
+    setIcon = () =>{
+        //string 으로 받은 icon 타입에 따라 출력
+        const {icon} = this.props;
+
+        if(icon === "warn") return "WARN";
+        else    return "WARN";
+    };
+
+    setButtons(){
+        const {btns, handleOnClick} = this.props;
+        
+        return btns.map((oneBtn, idx) => {
+            return (
+                <NoticeBtn
+                    key={`btn${idx}`}
+                    onClick={(event) => {
+                        this.setState({open: false}, () => handleOnClick(oneBtn.id, event))
+                }}>
+                    {oneBtn}
+                </NoticeBtn>
+            );
+        })
+    };
+
+    render(){
+        const {open} = this.state;
+        const {content} = this.props;
+    
+        return (
+            <>
+            {open ?
+                <Modal>
+                    <ModalCloser onClick={this.handleOnClose} />
+                    <Section>
+                        <ModalMain>
+                            <NoticeContainer>
+                                <Notice>
+                                    {/* <img src={warning} alt="warning" /> */}
+                                    <div>{this.setIcon()}</div>
+                                    <div>{content}</div>
+                                </Notice>
+                            </NoticeContainer>
+                        </ModalMain>
+                        <ModalFooter>
+                            {this.setButtons()}
+                        </ModalFooter>
+                    </Section>
+                </Modal>
+                :
+                null
+            }
+            </>
+        )
+    }
 };
-export default NoticeAlert;
 
+const NoticeBtn = mstyled(Button)({
+        flex: 1,
+        margin: "0 8px 0 8px",
+        borderRadius: "5px",
+        fontSize: "1em",
+        border: "1px solid #e5e8ec",
+        background: "#e5e8ec",
+});
+  
 const Modal = styled.div`
     position: fixed;
     display: -webkit-box;
@@ -124,3 +150,5 @@ const ModalCloser = styled.div`
     width: 100%;
     height: 100%;
 `;
+
+export default NoticeAlert;
